backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
paho-mqtt==2.1.0
websockets==15.0.1
python-dotenv==1.0.0
